pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "artx"
version = "0.5.0"
description = "Download high-resolution Google Arts & Culture images by stitching tiles"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "httpx>=0.28,<0.29",
    "pillow>=11.1.0,<12",
    "pycryptodome>=3.23.0,<4",
    "rich>=14.1.0,<15",
]

[project.scripts]
artx = "googleart_download.cli:main"

[project.optional-dependencies]
large-images = [
    "numpy>=2.4.3",
    "pyvips>=3.1.1",
    "tifffile>=2026.3.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/googleart_download"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.6",
]
